+++ webui/src/shared/api/index.ts
@@ -5,6 +5,7 @@ export * from "./history";
 export * from "./daemon";
 export * from "./moduleUpdate";
 export * from "./appUpdate";
+export * from "./silentUnlink";
 export * from "./batterySnapshot";
 export * from "./statusBundle";
 export * from "./events";